+++ docs/CHANGELOG.md
@@ -5,3 +5,4 @@ Changelog
 -------------------------
 - Init: Initialize Module "Text editor"
 - Fix #5: Position of "Edit" Handler
+- Enh #2: Create new text files